EES-3883 Move ThemeController.GetPublicationTree to PublicationController.GetPublicationTree
