+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api.Tests/Controllers/ThemeControllerTests.cs
@@ -5,7 +5,6 @@ using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Tests.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Content.Api.Controllers;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Moq;
 using Xunit;
@@ -17,23 +16,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Tests.Controlle
 {
     public class ThemeControllerTests
     {
-        private static readonly List<ThemeTree> Themes = new()
-        {
-            new()
-            {
-                Topics = new List<TopicTree>
-                {
-                    new()
-                    {
-                        Publications = new List<PublicationTreeNode>
-                        {
-                            new()
-                        }
-                    }
-                }
-            }
-        };
-
         private static readonly List<AllMethodologiesThemeViewModel> MethodologyThemes = new()
         {
             new AllMethodologiesThemeViewModel
@@ -66,32 +48,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Tests.Controlle
         };
 
         [Fact]
-        public async Task GetPublicationTree()
-        {
-            var themeCacheService = new Mock<IThemeCacheService>(Strict);
-
-            var controller = BuildController(themeCacheService: themeCacheService.Object);
-
-            themeCacheService
-                .Setup(s => s.GetPublicationTree(PublicationTreeFilter.FindStatistics))
-                .ReturnsAsync(Themes);
-
-            var result = await controller.GetPublicationTree(PublicationTreeFilter.FindStatistics);
-
-            VerifyAllMocks(themeCacheService);
-
-            var publicationTree = result.AssertOkResult();
-
-            var theme = Assert.Single(publicationTree);
-
-            Assert.IsType<ThemeTree>(theme);
-
-            var topic = Assert.Single(theme.Topics);
-
-            Assert.Single(topic.Publications);
-        }
-
-        [Fact]
         public async Task GetMethodologyThemes()
         {
             var methodologyCacheService = new Mock<IMethodologyCacheService>(Strict);
@@ -110,12 +66,10 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Tests.Controlle
         }
 
         private static ThemeController BuildController(
-            IMethodologyCacheService? methodologyCacheService = null,
-            IThemeCacheService? themeCacheService = null)
+            IMethodologyCacheService? methodologyCacheService = null)
         {
             return new ThemeController(
-                methodologyCacheService ?? Mock.Of<IMethodologyCacheService>(Strict),
-                themeCacheService ?? Mock.Of<IThemeCacheService>(Strict));
+                methodologyCacheService ?? Mock.Of<IMethodologyCacheService>(Strict));
         }
     }
 }
