+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/Controllers/PublicationController.cs
@@ -1,4 +1,5 @@
 ﻿#nullable enable
+using System.Collections.Generic;
 using System.Net.Mime;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
@@ -7,6 +8,7 @@ using GovUk.Education.ExploreEducationStatistics.Common.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Content.Api.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
@@ -26,6 +28,21 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Controllers
         {
             _publicationCacheService = publicationCacheService;
             _publicationService = publicationService;
+        }
+
+        [HttpGet("publication-tree")]
+        public async Task<ActionResult<IList<ThemeTree>>> GetPublicationTree(
+            [FromQuery(Name = "publicationFilter")]
+            PublicationTreeFilter? filter = null)
+        {
+            if (filter == null)
+            {
+                return new BadRequestResult();
+            }
+
+            return await _publicationCacheService
+                .GetPublicationTree(filter.Value)
+                .HandleFailuresOrOk();
         }
 
         [HttpGet("publications")]
