+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/Controllers/ThemeController.cs
@@ -4,7 +4,6 @@ using System.Net.Mime;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
@@ -15,28 +14,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Controllers
     public class ThemeController : ControllerBase
     {
         private readonly IMethodologyCacheService _methodologyCacheService;
-        private readonly IThemeCacheService _themeCacheService;
 
         public ThemeController(
-            IMethodologyCacheService methodologyCacheService,
-            IThemeCacheService themeCacheService)
+            IMethodologyCacheService methodologyCacheService)
         {
             _methodologyCacheService = methodologyCacheService;
-            _themeCacheService = themeCacheService;
-        }
-
-        [HttpGet("themes")]
-        public async Task<ActionResult<IList<ThemeTree>>> GetPublicationTree(
-            [FromQuery(Name = "publicationFilter")] PublicationTreeFilter? filter = null)
-        {
-            if (filter == null)
-            {
-                return new BadRequestResult();
-            }
-
-            return await _themeCacheService
-                .GetPublicationTree(filter.Value)
-                .HandleFailuresOrOk();
         }
 
         [HttpGet("methodology-themes")]
