--- src/GovUk.Education.ExploreEducationStatistics.Content.Services/Interfaces/Cache/IThemeCacheService.cs
@@ -1,17 +1,0 @@
-﻿#nullable enable
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using GovUk.Education.ExploreEducationStatistics.Common.Model;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Requests;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
-using Microsoft.AspNetCore.Mvc;
-
-namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
-
-public interface IThemeCacheService
-{
-    Task<Either<ActionResult, IList<ThemeTree>>> GetPublicationTree(
-        PublicationTreeFilter filter);
-    
-    Task<IList<ThemeTree>> UpdatePublicationTree();
-}
