+++ src/explore-education-statistics-admin/src/prototypes/components/PrototypeFilters3.tsx
@@ -5,7 +5,7 @@ import Button from '@common/components/Button';
 import ButtonText from '@common/components/ButtonText';
 import FormRadioGroup from '@admin/prototypes/components/PrototypeFormRadioGroup';
 import { releaseTypes } from '@common/services/types/releaseType';
-import { Theme } from '@common/services/themeService';
+import { Theme } from '@common/services/publicationService';
 import React from 'react';
 import Modal from '@common/components/Modal';
 import useToggle from '@common/hooks/useToggle';
