+++ src/explore-education-statistics-common/src/modules/table-tool/components/PublicationForm.tsx
@@ -10,7 +10,10 @@ import InsetText from '@common/components/InsetText';
 import SummaryList from '@common/components/SummaryList';
 import SummaryListItem from '@common/components/SummaryListItem';
 import WizardStepSummary from '@common/modules/table-tool/components/WizardStepSummary';
-import { PublicationSummary, Theme } from '@common/services/themeService';
+import {
+  PublicationTreeSummary,
+  Theme,
+} from '@common/services/publicationService';
 import Yup from '@common/validation/yup';
 import { InjectedWizardProps } from '@common/modules/table-tool/components/Wizard';
 import WizardStepFormActions from '@common/modules/table-tool/components/WizardStepFormActions';
@@ -25,7 +28,7 @@ export interface PublicationFormValues {
 }
 
 export type PublicationFormSubmitHandler = (values: {
-  publication: PublicationSummary;
+  publication: PublicationTreeSummary;
 }) => void;
 
 const formId = 'publicationForm';
