+++ src/explore-education-statistics-common/src/services/publicationService.ts
@@ -5,6 +5,7 @@ import {
   MethodologySummary,
   ExternalMethodology,
 } from '@common/services/types/methodology';
+import { PublicationType } from '@common/services/types/publicationType';
 import { Paging } from '@common/services/types/pagination';
 import { PartialDate } from '@common/utils/date/partialDate';
 import { contentApi } from './api';
@@ -111,8 +112,7 @@ interface PublicationListRequest {
 
 export interface Release<
   ContentBlockType extends ContentBlock = ContentBlock,
-  DataBlockType extends DataBlock = DataBlock,
-  PublicationType = Publication
+  DataBlockType extends DataBlock = DataBlock
 > {
   id: string;
   title: string;
@@ -126,7 +126,7 @@ export interface Release<
   keyStatisticsSecondarySection: ContentSection<DataBlockType>;
   headlinesSection: ContentSection<ContentBlockType>;
   relatedDashboardsSection?: ContentSection<ContentBlockType>; // optional because older releases may not have this section
-  publication: PublicationType;
+  publication: Publication;
   latestRelease: boolean;
   nextReleaseDate?: PartialDate;
   relatedInformation: BasicLink[];
@@ -160,6 +160,37 @@ export interface PublicationReleaseSummary extends ReleaseSummary {
 export interface PreReleaseAccessListSummary extends ReleaseSummary {
   publication: PublicationSummary;
   preReleaseAccessList: string;
+}
+
+export interface PublicationTreeSummary {
+  id: string;
+  title: string;
+  slug: string;
+  type: PublicationType;
+  legacyPublicationUrl?: string;
+  isSuperseded: boolean;
+}
+
+export interface Topic {
+  id: string;
+  title: string;
+  summary: string;
+  publications: PublicationTreeSummary[];
+}
+
+export interface Theme {
+  id: string;
+  title: string;
+  summary: string;
+  topics: Topic[];
+}
+
+interface PublicationTreeOptions {
+  publicationFilter?:
+    | 'FindStatistics'
+    | 'DataTables'
+    | 'DataCatalogue'
+    | 'FastTrack';
 }
 
 export default {
@@ -209,6 +240,13 @@ export default {
     return contentApi.get(
       `/publications/${publicationSlug}/releases/${releaseSlug}/prerelease-access-list`,
     );
+  },
+  getPublicationTree({
+    publicationFilter,
+  }: PublicationTreeOptions = {}): Promise<Theme[]> {
+    return contentApi.get('/publication-tree', {
+      params: { publicationFilter },
+    });
   },
   listPublications(
     params: PublicationListRequest,
