+++ src/explore-education-statistics-frontend/src/modules/find-statistics/FindStatisticsPageCurrent.tsx
@@ -1,7 +1,7 @@
 import AccordionSection from '@common/components/AccordionSection';
 import Details from '@common/components/Details';
 import RelatedInformation from '@common/components/RelatedInformation';
-import { Theme } from '@common/services/themeService';
+import { Theme } from '@common/services/publicationService';
 import Link from '@frontend/components/Link';
 import Page from '@frontend/components/Page';
 import PageSearchFormWithAnalytics from '@frontend/components/PageSearchFormWithAnalytics';
