+++ src/explore-education-statistics-frontend/src/modules/find-statistics/__tests__/__data__/testThemeData.ts
@@ -1,4 +1,4 @@
-import { Theme } from '@common/services/themeService';
+import { Theme } from '@common/services/publicationService';
 
 export const testThemes: Theme[] = [
   {
